:recycle: refactor(element): 폴더 생성 및 파일 경로 수정
